+++ include/elastic/binary_iserialize.hpp
@@ -99,8 +99,8 @@ namespace elastic
 		{
 			using typex = std::conditional_t<
 				std::is_enum_v<_Ty>, detail::identify_t<load_enum_type<_Archive>>,
-				std::conditional_t<optional_t<_Ty>, detail::identify_t<laod_optional_type<_Archive>>,
-								   std::conditional_t<unsign_t<_Ty> || fixed_t<_Ty>,
+				std::conditional_t<detail::optional_t<_Ty>, detail::identify_t<laod_optional_type<_Archive>>,
+								   std::conditional_t<detail::fixed_t<_Ty>,
 													  detail::identify_t<load_unsign_or_fixed_type<_Archive>>,
 													  detail::identify_t<load_non_pointer_type<_Archive>>>>>;
