+++ include/elastic/binary_oserialize.hpp
@@ -74,7 +74,7 @@ namespace elastic
 			template <typename _Ty>
 			static void invoke(_Archive& ar, _Ty&& t)
 			{
-				if constexpr (optional_t<std::remove_cvref_t<_Ty>>)
+				if constexpr (detail::optional_t<std::remove_cvref_t<_Ty>>)
 				{
 					ar << *t;
 				}
@@ -92,7 +92,7 @@ namespace elastic
 
 			using typex = std::conditional_t<
 				std::is_enum_v<type>, detail::identify_t<save_enum_type<_Archive>>,
-				std::conditional_t<attribute_t<type>, detail::identify_t<save_optional_type<_Archive>>,
+				std::conditional_t<detail::attribute_t<type>, detail::identify_t<save_optional_type<_Archive>>,
 								   detail::identify_t<save_non_pointer_type<_Archive>>>>;
 
 			typex::invoke(ar, std::forward<_Ty>(t));
