+++ include/elastic/detail/concepts.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "../types/fixed.hpp"
 #include "type_traits.hpp"
 
 #include <utility>
@@ -87,5 +88,25 @@ namespace elastic
 
 		template <typename _Ty>
 		concept string_t = std::is_same_v<_Ty, std::string>;
+
+		template <typename _Ty>
+		struct is_fixed : std::false_type
+		{};
+
+		template <typename _Ty>
+		struct is_fixed<fixed<_Ty>> : std::true_type
+		{};
+
+		template <typename _Ty>
+		concept fixed_t = is_fixed<_Ty>::value;
+
+		template <typename _Ty>
+		concept optional_t = requires(_Ty value) {
+								 value.has_value();
+								 *value;
+							 };
+
+		template <typename _Ty>
+		concept attribute_t = optional_t<std::remove_cvref_t<_Ty>> || fixed_t<std::remove_cvref_t<_Ty>>;
 	} // namespace detail
 } // namespace elastic
