+++ include/elastic/types/bytes.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <vector>
+
+namespace elastic
+{
+	using bytes = std::vector<std::byte>;
+}
