+++ include/elastic/types/optional.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include <optional>
+
+namespace elastic
+{
+	template <typename _Ty>
+	using optional = std::optional<_Ty>;
+}
