+++ include/elastic/types/repeated.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include <vector>
+
+namespace elastic
+{
+	template <typename _Ty>
+	using repeated = std::vector<_Ty>;
+}
