+++ include/elastic/types/string.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+
+namespace elastic
+{
+	using string = std::string;
+}
