+++ test/test.vcxproj
@@ -157,9 +157,13 @@
     <ClInclude Include="..\include\elastic\interface_archive.hpp" />
     <ClInclude Include="..\include\elastic\reflect.hpp" />
     <ClInclude Include="..\include\elastic\tuple_size.hpp" />
+    <ClInclude Include="..\include\elastic\types\bytes.hpp" />
+    <ClInclude Include="..\include\elastic\types\fixed.hpp" />
     <ClInclude Include="..\include\elastic\types\message.hpp" />
+    <ClInclude Include="..\include\elastic\types\optional.hpp" />
+    <ClInclude Include="..\include\elastic\types\repeated.hpp" />
     <ClInclude Include="..\include\elastic\types\sequence.hpp" />
-    <ClInclude Include="..\include\elastic\types\types.hpp" />
+    <ClInclude Include="..\include\elastic\types\string.hpp" />
     <ClInclude Include="..\include\elastic\types\varint.hpp" />
     <ClInclude Include="attr.h" />
     <ClInclude Include="exception.h" />
